Moved partition create task to maintenance task
