--- apiserver/src/main/java/org/dependencytrack/event/MetricsPartitionCreateEvent.java
@@ -1,29 +1,0 @@
-/*
- * This file is part of Dependency-Track.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * Copyright (c) OWASP Foundation. All Rights Reserved.
- */
-package org.dependencytrack.event;
-
-import alpine.event.framework.Event;
-
-/**
- * Defines an {@link Event} used to trigger creation of new partitions for the day.
- *
- * @since 4.6.0
- */
-public class MetricsPartitionCreateEvent implements Event {
-}
