+++ apiserver/src/main/java/org/dependencytrack/tasks/maintenance/MetricsMaintenanceTask.java
@@ -31,6 +31,7 @@ import java.time.Duration;
 import static net.javacrumbs.shedlock.core.LockAssert.assertLocked;
 import static org.dependencytrack.model.ConfigPropertyConstants.MAINTENANCE_METRICS_RETENTION_DAYS;
 import static org.dependencytrack.persistence.jdbi.JdbiFactory.openJdbiHandle;
+import static org.dependencytrack.persistence.jdbi.JdbiFactory.useJdbiHandle;
 import static org.dependencytrack.util.LockProvider.executeWithLock;
 import static org.dependencytrack.util.TaskUtil.getLockConfigForTask;
 
@@ -76,6 +77,9 @@ public class MetricsMaintenanceTask implements Subscriber {
     private Statistics informLocked(final Handle jdbiHandle) {
         assertLocked();
 
+        // Create new partitions for today.
+        useJdbiHandle(handle -> handle.attach(MetricsDao.class).createMetricsPartitionsForToday());
+
         final var configPropertyDao = jdbiHandle.attach(ConfigPropertyDao.class);
         final var metricsDao = jdbiHandle.attach(MetricsDao.class);
 
@@ -88,5 +92,4 @@ public class MetricsMaintenanceTask implements Subscriber {
 
         return new Statistics(retentionDuration, numDeletedComponent, numDeletedProject, numDeletedPortfolio);
     }
-
 }
