+++ apiserver/src/main/resources/application.properties
@@ -1351,27 +1351,10 @@ task.component.metadata.maintenance.lock.min.duration=PT1M
 # @required
 task.metrics.partition.create.cron=0 12 * * *
 
-# Maximum duration in ISO 8601 format for which the metrics maintenance task will hold a lock.
-# <br/><br/>
-# The duration should be long enough to cover the task's execution duration.
-#
-# @category: Task Scheduling
-# @type:     duration
-# @required
-task.metrics.partition.create.lock.max.duration=PT15M
-
-# Minimum duration in ISO 8601 format for which the metrics maintenance task will hold a lock.
-# <br/><br/>
-# The duration should be long enough to cover eventual clock skew across API server instances.
-#
-# @category: Task Scheduling
-# @type:     duration
-# @required
-task.metrics.partition.create.lock.min.duration=PT1M
-
 # Cron expression of the metrics maintenance task.
 # <br/><br/>
-# The task deletes records older than the configured metrics retention duration from the following tables:
+# The task creates new partitions for the day for the following tables
+# And deletes records older than the configured metrics retention duration from the following tables:
 # <ul>
 #   <li><code>DEPENDENCYMETRICS</code></li>
 #   <li><code>PROJECTMETRICS</code></li>
@@ -1381,7 +1364,7 @@ task.metrics.partition.create.lock.min.duration=PT1M
 # @category: Task Scheduling
 # @type:     cron
 # @required
-task.metrics.maintenance.cron=0 */3 * * *
+task.metrics.maintenance.cron=0 12 * * *
 
 # Maximum duration in ISO 8601 format for which the metrics maintenance task will hold a lock.
 # <br/><br/>
